move test in tests/test_cli.py
