+++ tests/test_cli.py
@@ -1,5 +1,6 @@
 from click.testing import CliRunner
 from unittest import mock
+import json
 import os
 import pytest
 import shutil
@@ -9,13 +10,19 @@ import subprocess
 
 from urllib.request import url2pathname
 from urllib.parse import urlparse, unquote
+import numpy as np
+import pandas as pd
 
+import mlflow
 from mlflow.cli import server, ui
+from mlflow import pyfunc
 from mlflow.server import handlers
 from mlflow.store.tracking.sqlalchemy_store import SqlAlchemyStore
 from mlflow.store.tracking.file_store import FileStore
 from mlflow.exceptions import MlflowException
 from mlflow.entities import ViewType
+
+from tests import pyfunc_serve_and_score_model
 
 
 def test_server_static_prefix_validation():
@@ -162,3 +169,27 @@ def test_mlflow_gc_not_deleted_run(file_store):
         )
     runs = store.search_runs(experiment_ids=["0"], filter_string="", run_view_type=ViewType.ALL)
     assert len(runs) == 1
+
+
+def test_mlflow_models_serve():
+    class MyModel(pyfunc.PythonModel):
+        def predict(self, context, model_input):  # pylint: disable=unused-variable
+            return np.array([1, 2, 3])
+
+    model = MyModel()
+
+    with mlflow.start_run():
+        mlflow.pyfunc.log_model(artifact_path="model", python_model=model)
+        model_uri = mlflow.get_artifact_uri("model")
+
+    data = pd.DataFrame({"a": [0]})
+
+    scoring_response = pyfunc_serve_and_score_model(
+        model_uri=model_uri,
+        data=data,
+        content_type=pyfunc.scoring_server.CONTENT_TYPE_JSON_SPLIT_ORIENTED,
+        extra_args=["--no-conda"],
+    )
+    assert scoring_response.status_code == 200
+    served_model_preds = np.array(json.loads(scoring_response.content))
+    np.testing.assert_array_equal(served_model_preds, model.predict(data, None))
